Validating UUID

+++ src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactRepository = require('../repositories/ContactRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -10,6 +11,11 @@ class ContactController {
 
   async show(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' });
+    }
+
     const contact = await ContactRepository.findById(id);
 
     if (!contact) {
@@ -28,14 +34,23 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' });
+    }
 
-    if (contactExists) {
-      return response.status(400).json({ error: 'This e-mail already in use' });
+    if (email) {
+      const contactExists = await ContactRepository.findByEmail(email);
+
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail already in use' });
+      }
     }
 
     const contact = await ContactRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.status(201).json(contact);
@@ -47,22 +62,36 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExists = await ContactRepository.findById(id);
-    if (!contactExists) {
-      return response.status(400).json({ error: 'Contact not found' });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactByEmail = await ContactRepository.findByEmail(email);
-    if (contactByEmail && contactByEmail.id === id) {
-      return response.status(400).json({ error: 'This email already in use' });
+    const contactExists = await ContactRepository.findById(id);
+
+    if (!contactExists) {
+      return response.status(400).json({ error: 'Contact not found' });
+    }
+
+    if (email) {
+      const contactByEmail = await ContactRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id === id) {
+        return response.status(400).json({ error: 'This email already in use' });
+      }
     }
 
     const contact = await ContactRepository.update(id, {
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -70,6 +99,10 @@ class ContactController {
 
   async delete(request, response) {
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid user id' });
+    }
 
     await ContactRepository.delete(id);
